requirements.txt
Flask==2.0.3
pandas==1.3.5
psycopg2==2.9.3
requests==2.26.0
python-dotenv==0.19.2